CONFIGDB.API: Organize code. Create package for normconfigs windows.
