+++ pyqt-apps/siriushla/as_ap_configdb/__init__.py
@@ -1,4 +1,4 @@
-from .ConfigManagerWindow import ConfigManagerWindow
+from .client_configdb import ConfigurationManager
 from .configdialogs import LoadConfigDialog, SaveConfigDialog, \
     RenameConfigDialog
 from .configwidgets import ConfigTableWidget, ConfigDBSearchEngine, \
