+++ pyqt-apps/siriushla/as_ap_configdb/normconfigs/TuneDlg.py
