+++ pyqt-apps/siriushla/as_ap_configdb/normconfigs/__init__.py
@@ -1,0 +1,1 @@
+from .ConfigManagerWindow import ConfigManagerWindow
